+++ src/Smartstore.Core/Catalog/Attributes/Domain/ProductVariantAttributeCombination.cs
@@ -92,6 +92,11 @@ namespace Smartstore.Core.Catalog.Attributes
         public decimal? Price { get; set; }
 
         /// <summary>
+        /// Gets or sets the compare price.
+        /// </summary>
+        public decimal? ComparePrice { get; set; }
+
+        /// <summary>
         /// Gets or sets the weight.
         /// </summary>
         public decimal? Weight { get; set; }